Chore: refactor shape restoration
