+++ src/careamics/lightning/dataset_ng/prediction/__init__.py
@@ -13,6 +13,6 @@ from .convert_prediction import (
     combine_samples,
     convert_prediction,
     decollate_image_region_data,
-    restore_original_shape,
 )
+from .restore_original_shape import restore_original_shape
 from .stitch_prediction import stitch_prediction, stitch_single_prediction
